day 05: try to use a "generation" to ignore events
If a crate was just moving, and will also be moved next, then its old
animationend event seems to still be queued when its new onanimationend
handler is added (and thus instantly fires).

This was an attempt to bind a monotonically increasing "generation" to
the handler, but of course that doesn't help -- the problem is the
event firing, not the handler.  The extra event just gets directed *to*
the correct handler.

This is still weird, because the triggered reflow *should* have
completely stopped the old animation on that element.