day 16: fix a couple bugs
Actually open the valve if we're staying put and opening the valve.

And allow us to stay in the same location if we've already opened the
valve -- this fixes the "no states!?" panic, but unfortunately it seems
to add states very widely.  Logically, we shouldn't be adding any new
unique paths after the 20th-or-so step, because the example input only
has ten valves.  After we've opened and visited every one of them, there
shouldn't be anything left...