day 16: recursively DFS for the maximal path
This is also something I found from the aforementioned Reddit thread,
but ... I seem to have an implementation detail wrong since it now
returns a wildly high value for the example data.

It stands to reason that this avoids recomputing the released pressure
for the same path prefixes over and over again.